notes/references/python_live_transcription/requirements.txt
google-cloud-speech>=2.21.0
pyaudio>=0.2.14
